+++ ts/packages/knowPro/src/hybrid/index.ts
@@ -2,7 +2,7 @@
 // Licensed under the MIT License.
 
 /**
- * Hybrid search is EXPERIMENTAL CODE
+ * Hybrid search is EXPERIMENTAL CODE and RAPIDLY EVOLVING
  *
  * It allows Structured RAG to use both its standard indexes (entity, action, topic)
  * and custom data frames in a query
@@ -10,4 +10,5 @@
 
 export * from "./dataFrame.js";
 export * from "./dataFrameQuery.js";
+export * from "./hybridConversation.js";
 export * as lang from "./dataFrameLangSearch.js";
